1.5. Testy, diagramy, dokumentacja i *gwiazdki
RhymersDemo:
	rozbicie metody main. Stworzenie nowej metody testRhymers.

Node: przeniesienie klasy do klasy IntLinkedList.
