--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,31 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class Node {
-
-    private int value;
-    private Node prev, next;
-
-    public Node(int i) {
-        value = i;
-    }
-
-    public Node getPrev() {
-        return prev;
-    }
-
-    public void setPrev(Node prev) {
-        this.prev = prev;
-    }
-
-    public Node getNext() {
-        return next;
-    }
-
-    public void setNext(Node next) {
-        this.next = next;
-    }
-
-    public int getValue() {
-        return value;
-    }
-}
